Update TunnelManager.py
Added a console with command processing, and set up the script to run normally if the console is not invoked. Also added threading so that the process monitor can run in the console without hanging up the prompt. The threaded class is not called when the console is not invoked, otherwise the script would die.