Update TunnelManager.py
Lowered the default refresh time to 1 minute, set path for log to /Applications/TunnelManager/TunnelManager.log

Eventually this will go back to a universal approach, but for the time being we are building for OSX machines only. Though this will run and install on Linux if you use the installer. Or simply create an /Applications directory on your linux machine and drop TunnelManager in there.