Update TunnelManager.py
Added a simple UI, with a "Connect" button to initiate the connection. Once connected, the button text changes to "Connected". There is currently no additional functionality built into the UI, as the UI is not the primary method of controlling the service; that would be the console. However, more functionality will be added soon so that those who prefer UI over console can have more control over the service.